+++ src/elements/db/SubmissionQuery.php
@@ -61,6 +61,9 @@ class SubmissionQuery extends ElementQuery
             'simpleform_submissions.data',
             'simpleform_submissions.userId',
             'simpleform_submissions.readStatus',
+            'simpleform_submissions.paymentStatus',
+            'simpleform_submissions.paymentAmount',
+            'simpleform_submissions.orderId',
         ]);
 
         if ($this->formId !== null) {